End com refactor
