+++ libbsn/src/communication/TCPReceive.cpp
@@ -1,4 +1,4 @@
-#include "../include/TCPReceive.hpp"
+#include "bsn/communication/TCPReceive.hpp"
 
 using namespace std;
