+++ libbsn/src/communication/TCPSend.cpp
@@ -1,4 +1,4 @@
-#include "../include/TCPSend.hpp"
+#include "bsn/communication/TCPSend.hpp"
 
 using namespace std;
 using namespace odcore;
